+++ backend/src/routes/index.ts
@@ -1,5 +1,5 @@
 import fastify, { FastifyInstance } from "fastify";
-import localGameRoute from './localgame'
+import localGameRoute from './local.game'
 
 const routes = async (fastify: FastifyInstance) => {
     await fastify.register(localGameRoute, {prefix: '/localgame'})
